Sync: Disallow deletes via sync by default.Due to a very nasty bug in Android 12   some phones started to delete contacts via Microsoft ActiveSync. Therefore we've   implemented a new setting in Group-Office that disallows deletes via sync by   default now. More info on the bug can be found here:   https://eu.community.samsung.com/t5/galaxy-z-fold-z-flip/outlook-contacts-gone-after-android-12-update/td-p/4567744/page/2

+++ www/modules/z-push/backend/go/goNote.php
@@ -9,6 +9,12 @@ use go\modules\community\notes\model\Note;
 class goNote extends GoBaseBackendDiff {
 
 	public function DeleteMessage($folderid, $id, $contentparameters) {
+
+		if(!go()->getAuthState()->getUser(['syncSettings'])->syncSettings->allowDeletes) {
+			ZLog::Write(LOGLEVEL_DEBUG, 'Deleting by sync is disabled in user settings');
+			throw new StatusException(SYNC_ITEMOPERATIONSSTATUS_DL_ACCESSDENIED);
+		}
+
 		try {
 			ZLog::Write(LOGLEVEL_DEBUG, 'goNote->DeleteMessage(' . $folderid . ',' . $id . ')');
 			$note = Note::findById($id);
